matplotlib tests in right place
